+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Initial release written in Swift 3.1
 * Integrate with Travis CI [#1](https://github.com/carousell/pickle/pull/1)
 * Support 3D touch photo preview [#3](https://github.com/carousell/pickle/pull/3)
+* Show the camera button in the center when the photo gallery is empty [#8](https://github.com/carousell/pickle/pull/8)